requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
sentence-transformers
torch
python-multipart